tui: split menu renderers and mouse handling out of menu.go (plan 023 F1 follow-up)
Move-only follow-up to C18 satisfying §7-F's ~450-line bar: menu.go
(803→314) keeps types, item builders, keyboard routing, and activation;
menu_render.go (345) takes the bar/dropdown renderers and overlay
splice; menu_mouse.go (155) takes handleMenuMouse/handleMenuMotion.
Zero logic edits — pure extractions verified by git diff -w.

Verified: make lint && make test && make test-race && make build.
